+++ packages/tools/pluggable-widgets-tools/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   We added support for association properties linked to a data source, introduced in Mendix 9.17.
+
 ## [9.13.4] - 2022-08-26
 
 ### Changed